feat: dynamic gui gen
