+++ src/cala/gui/routes.py
@@ -1,22 +1,22 @@
 import base64
-from datetime import datetime
-from queue import Queue
-from typing import Annotated
 
 import yaml
-from fastapi import APIRouter, WebSocket, BackgroundTasks, HTTPException, UploadFile, Header
-from fastapi.encoders import jsonable_encoder
+from fastapi import APIRouter, BackgroundTasks, HTTPException, UploadFile
 from fastapi.requests import Request
-from fastapi.responses import FileResponse, HTMLResponse
+from fastapi.responses import HTMLResponse
 from fastapi.templating import Jinja2Templates
 from noob import SynchronousRunner
 from noob.tube import TubeSpecification, Tube
-from starlette.responses import JSONResponse
+from starlette.responses import FileResponse
+from starlette.websockets import WebSocket
 
 from cala.config import config
 from cala.gui.const import TEMPLATES_DIR
 from cala.gui.deps import Spec
+from cala.gui.util import QManager
+from cala.logging import init_logger
 
+logger = init_logger(__name__)
 templates = Jinja2Templates(directory=TEMPLATES_DIR)
 
 router = APIRouter()
@@ -26,30 +26,6 @@ _thread = None
 _tube_config = None
 
 
-class QManager:
-    _qs: dict[str, Queue] = {}
-
-    @classmethod
-    def get_queue(cls, name: str) -> Queue:
-        if cls._qs.get(name) is None:
-            cls._qs[name] = Queue()
-        return cls._qs.get(name)
-
-
-@router.get("/now", response_class=HTMLResponse)
-async def metadata(request: Request):
-    return templates.TemplateResponse(request, "includes/metadata.html", {"now": datetime.now()})
-
-
-@router.get("/open-time", response_class=HTMLResponse)
-async def list_todos(request: Request, hx_request: Annotated[str | None, Header()] = None):
-    if hx_request:
-        return templates.TemplateResponse(
-            request=request, name="includes/metadata.html", context={"open_time": datetime.now()}
-        )
-    return JSONResponse(content=jsonable_encoder(datetime.now()))
-
-
 @router.get("/", response_class=HTMLResponse)
 async def get(request: Request, spec: Spec) -> HTMLResponse:
     response = templates.TemplateResponse(request, "index.html", {"config": spec.model_dump()})
@@ -57,32 +33,6 @@ async def get(request: Request, spec: Spec) -> HTMLResponse:
     response.set_cookie(key="config", value=config, samesite="lax")
 
     return response
-
-
-@router.websocket("/ws")
-async def websocket_endpoint(websocket: WebSocket) -> None:
-    """Handle WebSocket connection and run streamers"""
-    q = QManager.get_queue("lineplot")
-
-    await websocket.accept()
-    # some kind of waiting mechanism here
-    # to keep the connection open until the client disconnects
-    while True:
-        event = q.get()
-        if event is None:
-            break
-        await websocket.send_json(event)
-
-
-@router.get("/{node_id}/stream.m3u8")
-async def stream(node_id: str, filename: str) -> FileResponse:
-    """Serve HLS stream files"""
-    stream_path = config.runtime_dir / node_id / filename
-    if stream_path.exists():
-        return FileResponse(str(stream_path))
-    else:
-        return HTMLResponse("File not found")
-    raise HTTPException(404, detail={"AppStreamError": f"Playlist not found: {stream_path}"})
 
 
 @router.post("/start")
@@ -97,8 +47,9 @@ def start(background: BackgroundTasks, gui_spec: Spec, request: Request) -> HTML
         runner = SynchronousRunner(tube=tube)
 
         def _cb(event):
-            q = QManager.get_queue("lineplot")
-            if event.condition == "what i want":
+            q = QManager.get_queue("prepped")
+            logger.warning(msg=event)
+            if event["node_id"] == "glow":
                 q.put(event)
 
         runner.add_callback(_cb)
@@ -130,6 +81,42 @@ def submit_tube(file: UploadFile, request: Request) -> HTMLResponse:
         raise HTTPException(422, f"Failed to load Tube specification. {e}")
     _tube_config = tube_config
     return HTMLResponse(f"Loaded: {tube_config["noob_id"]}")
-    # return templates.TemplateResponse(
-    #     request, "partials/tube-config.html", {"tube_config": tube_config}
-    # )
+
+
+@router.post("/player/{node_id}")
+async def player(node_id: str, request: Request) -> HTMLResponse:
+    """Serve video player DOM"""
+    stream_path = config.runtime_dir / node_id / "stream.m3u8"
+    print(f"stream_path={stream_path}")
+    if stream_path.exists():  # later change this to "has chunk ext files"
+        return templates.TemplateResponse(
+            request, "partials/player.html", {"id": node_id, "path": stream_path}
+        )
+    else:
+        raise HTTPException(404)
+
+
+@router.get("/stream/{node_id}")
+async def stream(node_id: str) -> FileResponse:
+    """Serve HLS stream files"""
+    stream_path = config.runtime_dir / node_id / "stream.m3u8"
+    print(f"stream_path={stream_path}")
+    if stream_path.exists():
+        return FileResponse(str(stream_path))
+    else:
+        return HTMLResponse("File not found")
+
+
+@router.websocket("/ws")
+async def websocket_endpoint(websocket: WebSocket) -> None:
+    """Handle WebSocket connection and run streamers"""
+    q = QManager.get_queue("lineplot")
+
+    await websocket.accept()
+    # some kind of waiting mechanism here
+    # to keep the connection open until the client disconnects
+    while True:
+        event = q.get()
+        if event is None:
+            break
+        await websocket.send_json(event)
