--- src/cala/gui/templates/includes/grid.html
@@ -1,17 +1,0 @@
-<div id="grids" class="grids">
-    <div class="video-container gfg">
-        <h3>{{ spec.grids['prep_movie'].id }}</h3>
-        <video
-                id={{ spec.grids[
-        "prep_movie"].id }}
-        class="video-js"
-        data-setup='{"fluid": true}'
-        controls
-        >
-        <source
-                src="{{ spec.grids['prep_movie'].id }}/stream.m3u8"
-                type="application/x-mpegURL"
-        />
-        </video>
-    </div>
-</div>
