+++ src/cala/gui/templates/includes/start.html
@@ -4,6 +4,6 @@
         hx-trigger="click"
         hx-post="/start"
         hx-swap="innerHTML"
-        hx-target="#grids"
+        hx-target="#main"
 >Start
 </button>
