+++ src/cala/gui/templates/index.html
@@ -12,7 +12,7 @@
     {% include "includes/filepicker.html" %}
     {% include "includes/start.html" %}
 </div>
-<div id="grids" class="grids"></div>
+<div id="main" class="main"></div>
 <script src="../static/main.js"></script>
 </body>
 </html>
