+++ src/cala/gui/templates/partials/player.html
@@ -1,0 +1,11 @@
+<video
+        id="{{ id }}"
+        class="video-js"
+        controls
+        data-setup='{"fluid": true}'
+>
+    <source
+            src="{{ path }}"
+            type="application/x-mpegURL"
+    />
+</video>
